feat: 회원정보 수정, 삭제 완료

+++ backend/src/main/java/com/ssafy/ododok/db/model/UserSurvey.java
@@ -18,7 +18,7 @@ public class UserSurvey {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int userSurveyId;
+    private Long userSurveyId;
 
     @OneToOne
     @JoinColumn(name="userId")
